refactor(bridge): replace LinkWithPrefixNetwork with AddressLink component
